Major changes in Particle ParticleSystem and Path.
The Path is now buuild on particles so the pathpoints get affected by the repellers.
Particles have now a affection boolean and a hidden boolean.
if affection is true. Particle gets affected
if hidden is true particle will not be displayed. This is used to store all particles, also the path particeles in the Arraylist ptclsList.